requirements.txt
flask
flask-cors
openai
python-docx
pdfplumber
gunicorn